+++ attributes/default.rb
@@ -62,6 +62,15 @@ default['openstack']['dashboard']['secret_key_content'] = nil
 
 default['openstack']['dashboard']['webroot'] = '/'
 
+# Dashboard specific database packages
+# Put common ones here and platform specific ones below.
+default['openstack']['dashboard']['db_python_packages'] = {
+  db2: [],
+  mysql: [],
+  postgresql: [],
+  sqlite: []
+}
+
 case node['platform_family']
 when 'rhel'
   default['openstack']['dashboard']['horizon_user'] = 'apache'
@@ -73,7 +82,7 @@ when 'rhel'
   default['openstack']['dashboard']['login_url'] = "#{node['openstack']['dashboard']['webroot']}auth/login/"
   default['openstack']['dashboard']['logout_url'] = "#{node['openstack']['dashboard']['webroot']}auth/logout/"
   default['openstack']['dashboard']['login_redirect_url'] = node['openstack']['dashboard']['webroot']
-  default['openstack']['db']['python_packages']['db2'].push 'python-ibm-db-django'
+  default['openstack']['dashboard']['db_python_packages']['db2'] = ['python-ibm-db-django']
   default['openstack']['dashboard']['platform'] = {
     'horizon_packages' => ['openstack-dashboard'],
     'memcache_python_packages' => ['python-memcached'],