+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is used to list changes made in each version of the openstack-dashboar
 * Sync conf files with Juno
 * Add optional section support for local_settings template
 * Update local_settings from 0644 to 0640
+* Fix python-ibm-db-django package polluting common package attribute
 
 ## 9.1
 * python_packages database client attributes have been moved to the -common cookbook